refactor: move test to compliance suite
